Okay, finished revising this
Got all the English revised by my Editor, might just do another review later on this week for consistency of style. As it is, each line is accurate and correct. This has taken me years of study, a big hobby of mine, and I decided to share all my notes from all the languages I had them into English for ease of access for more people around the world. Please feel free to use this information for recreational purposes only. Although the I-Ching has been a guide for thousands of years in China, don't let this rule your life choices, instead, meditate on each statement and find its meaning for you in that particular moment. It is recommended to get one interpretation per day, and try to read it throughout the whole day and see its significance with the events that take place in your life that very same day.